Map double->int data

+++ morphac/environment/binding/src/map_binding.cc
@@ -6,7 +6,7 @@ namespace binding {
 
 namespace py = pybind11;
 
-using Eigen::MatrixXd;
+using Eigen::MatrixXi;
 using morphac::environment::Map;
 
 void define_map_binding(py::module& m) {
@@ -14,7 +14,7 @@ void define_map_binding(py::module& m) {
 
   map.def(py::init<const double, const double, const double>(),
           py::arg("width"), py::arg("height"), py::arg("resolution"));
-  map.def(py::init<const MatrixXd&, const double>(), py::arg("data"),
+  map.def(py::init<const MatrixXi&, const double>(), py::arg("data"),
           py::arg("resolution"));
   map.def_property_readonly("width", &Map::get_width);
   map.def_property_readonly("height", &Map::get_height);
