+++ morphac/environment/src/map.cc
@@ -3,7 +3,7 @@
 namespace morphac {
 namespace environment {
 
-using Eigen::MatrixXd;
+using Eigen::MatrixXi;
 
 using morphac::environment::Map;
 
@@ -30,10 +30,10 @@ Map::Map(const double width, const double height, const double resolution)
                     std::numeric_limits<double>::epsilon(),
                 std::invalid_argument, "Invalid resolution.");
 
-  data_ = MatrixXd::Zero(rows, cols);
+  data_ = MatrixXi::Zero(rows, cols);
 }
 
-Map::Map(const MatrixXd& data, const double resolution)
+Map::Map(const MatrixXi& data, const double resolution)
     : width_(data.cols() * resolution),
       height_(data.rows() * resolution),
       resolution_(resolution) {
@@ -68,11 +68,11 @@ double Map::get_height() const { return height_; }
 
 double Map::get_resolution() const { return resolution_; }
 
-const MatrixXd& Map::get_data() const { return data_; }
+const MatrixXi& Map::get_data() const { return data_; }
 
-MatrixXd& Map::get_data_ref() { return data_; }
+MatrixXi& Map::get_data_ref() { return data_; }
 
-void Map::set_data(const MatrixXd& data) {
+void Map::set_data(const MatrixXi& data) {
   // The data needs to have the same dimensions
   MORPH_REQUIRE((data.rows() * this->resolution_) == this->height_,
                 std::invalid_argument, "Data height does not match.");
@@ -81,7 +81,7 @@ void Map::set_data(const MatrixXd& data) {
   data_ = data;
 }
 
-Map Map::Evolve(const MatrixXd& data) {
+Map Map::Evolve(const MatrixXi& data) {
   MORPH_REQUIRE(this->data_.rows() == data.rows(), std::invalid_argument,
                 "Data height does not match.")
   MORPH_REQUIRE(this->data_.cols() == data.cols(), std::invalid_argument,
